// src/models/movie.ts
export interface IMovie {
  adult: boolean;
  backdrop_path: string | null;
  genre_ids: number[];
  id: number;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string | null;
  release_date: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
}

export interface IDetailedMovie {
  adult: boolean;
  backdrop_path: string;
  budget: number;
  genres: TGanres[];
  id: number;
  original_language: string;
  original_title: string;
  overview: string;
  poster_path: string | null;
  release_date: string;
  runtime: number;
  tagline: string | null;
  title: string;
  video: boolean;
  vote_average: number;
}

type TGanres = {
  id: string;
  name: string;
};